+++ src/Symfony/Bundle/FrameworkBundle/Tests/DependencyInjection/Fixtures/php/request_response_serializer.php
@@ -1,0 +1,10 @@
+<?php
+
+$container->loadFromExtension('framework', [
+    'request' => [
+        'serializer' => 'serializer.request',
+    ],
+    'response' => [
+        'serializer' => 'serializer.response',
+    ],
+]);